pyproject.toml
[tool.poetry]
name = "sandbox"
version = "0.1.0"
description = "just a sandbox"
authors = ["Bruno Fontana <fontanads@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.23.5"
pandas = "^1.5.3"
altair = "^4.2.2"
jupyterlab = "^3.6.2"
prophet = "^1.1.2"
plotly = "^5.13.0"
matplotlib = "^3.6.3"
ipywidgets = "^8.0.4"
google-api-python-client = "^2.82.0"
streamlit = "^1.20.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-auth = "^2.16.3"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.19.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
tox = "^4.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
